// frontend/src/app/features/avaliacoes/models/avaliacao.model.ts
export interface AvaliacaoRequestDTO {
  dataAvaliacao: string;

  remadaBracoD: number;
  remadaBracoE: number;

  elevacaoLatD: number;
  elevacaoLatE: number;

  extensaoJoelhoD: number;
  extensaoJoelhoE: number;

  flexaoJoelhoD: number;
  flexaoJoelhoE: number;

  extensaoQuadrilD: number;
  extensaoQuadrilE: number;
}